pyproject.toml
[project]
name = "embgen"
version = "0.1.0"
description = "Embedded code generator - Generate code from YAML definitions"
requires-python = ">=3.13"
dependencies = [
    "jinja2>=3.1.6",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "rich>=14.2.0",
]

[project.scripts]
embgen = "embgen.cli:main"

[dependency-groups]
dev = ["pre-commit>=4.5.1", "pytest>=9.0.2", "ruff>=0.14.10", "ty>=0.0.5"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
# Force sequential execution (disable any parallel plugins)
addopts = [
    "--tb=short",       # Shorter traceback format
    "--strict-markers", # Require markers to be defined
]
# Ensure no parallel execution
testpaths = ["test"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
